# api-gateway/api/docs/swagger.yaml
basePath: /v1
definitions:
  models.AuthResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      passward:
        type: string
      phone_number:
        type: string
      token:
        type: string
      type:
        type: string
      user_name:
        type: string
    type: object
...
  models.CustomerAuthResponse:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
  models.CustomerRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      passward:
        type: string
      phone_number:
        type: string
      type:
        enum:
        - superadmin
        - user
        type: string
      user_name:
        type: string
    required:
    - type
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  models.GetListSalonsResponse:
    properties:
      count:
        type: integer
      salons:
        items:
          $ref: '#/definitions/models.Salon'
        type: array
    type: object
...
  models.LogInRequest:
    properties:
      email:
        type: string
      passward:
        type: string
    type: object
  models.ResponseOK:
    properties:
      message:
        type: string
    type: object
  models.Salon:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      end_time:
        type: string
      id:
        type: string
      image_url:
        type: string
      latitude:
        type: string
      longitude:
        type: string
      name:
        type: string
      phone_number:
        type: string
      rating:
        type: integer
      start_time:
        type: string
      updated_at:
        type: string
    type: object
  models.SalonRequest:
    properties:
      address:
        type: string
      email:
        type: string
      end_time:
        type: string
      image_url:
        type: string
      latitude:
        type: string
      longitude:
        type: string
      name:
        type: string
      phone_number:
        type: string
      rating:
        type: integer
      start_time:
        type: string
    type: object
...
  models.UpdatePasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  models.VerifyRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
host: localhost:9090
info:
  contact: {}
  title: Swagger for barber shop api
  version: "1.0"
paths:
...
  /customer/forgot-password:
    post:
      consumes:
      - application/json
      description: This api for forgot password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: forgot password
      tags:
      - customer_auth
...
  /customer/login:
    post:
      consumes:
      - application/json
      description: This api for login customer
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.LogInRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CustomerAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login customer
      tags:
      - customer_auth
...
  /customer/register:
    post:
      consumes:
      - application/json
      description: This api for registering a customer
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: register a customer
      tags:
      - customer_auth
  /customer/update-password:
    post:
      consumes:
      - application/json
      description: This api for updating customer password
      parameters:
      - description: Password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: update password
      tags:
      - customer_auth
...
  /customer/verify:
    post:
      consumes:
      - application/json
      description: This api for verification customer
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VerifyRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify customer
      tags:
      - customer_auth
  /customer/verify-forgot-password:
    post:
      consumes:
      - application/json
      description: Verify forgot password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify forgot password
      tags:
      - customer_auth
...
  /salon/create:
    post:
      consumes:
      - application/json
      description: This api for creating salon
      parameters:
      - description: Salon
        in: body
        name: salon
        required: true
        schema:
          $ref: '#/definitions/models.SalonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Salon'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a salon
      tags:
      - salon
  /salon/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This api for deleting salon by id
      parameters:
      - description: SalonID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete salon by id
      tags:
      - salon
  /salon/get/{id}:
    get:
      consumes:
      - application/json
      description: This api for getting salon by id
      parameters:
      - description: SalonID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Salon'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get salon by id
      tags:
      - salon
  /salon/update/{id}:
    put:
      consumes:
      - application/json
      description: This api for updating salon
      parameters:
      - description: SalonID
        in: path
        name: id
        required: true
        type: string
      - description: Salon
        in: body
        name: salon
        required: true
        schema:
          $ref: '#/definitions/models.SalonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Salon'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update salon
      tags:
      - salon
  /salons/list:
    get:
      consumes:
      - application/json
      description: This api for getting list of salons
      parameters:
      - default: 10
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        in: query
        name: page
        required: true
        type: integer
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetListSalonsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: get list salons
      tags:
      - salon
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"